Moved analytics to their own custom script
