+++ Assets/ScoreReporter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1e032a9faebb1544bcf75437dbc1740
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
