+++ Assets/Scripts/DifficultyRegulator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Analytics;
+
 
 public class DifficultyRegulator : MonoBehaviour {
 
@@ -39,27 +39,12 @@ public class DifficultyRegulator : MonoBehaviour {
 
 	public int BreakTimeSeconds = 10;
 
-    private Dictionary<string, object> parameters
-         = new Dictionary<string, object>();
 
-
-    void GameOverListener() {
-        parameters["Score"] = ScoreDisplay.Score;
-        AnalyticsEvent.Custom("HighScore",parameters);
-    }
-
-    private void OnDestroy()
-    {
-        if (GameEventSystem.Instance != null) {
-            GameEventSystem.Instance.GameEnded.RemoveListener(GameOverListener);
-                                                    }
-    }
 
     // Use this for initialization
     void Start () {
 		this.Spawners = GameObject.Find("Bars").GetComponentsInChildren<PatronSpawner>();
-        GameEventSystem.Instance.GameEnded.AddListener(GameOverListener);
-        parameters.Add("Score", 0);
+
         this.ScoreDisplay.Bucks = this.StartingBucks;
 	}
